Mega Refactoring

+++ src/main/java/com/programm/vertx/response/Pagination.java
@@ -1,25 +1,25 @@
 package com.programm.vertx.response;
 
 public class Pagination {
-    private final int total;
-    private final int skip;
-    private final int perPage;
+    private final long total;
+    private final long skip;
+    private final long perPage;
 
-    public Pagination(int total, int skip, int perPage) {
+    public Pagination(long total, long skip, long perPage) {
         this.total = total;
         this.skip = skip;
         this.perPage = perPage;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public int getSkip() {
+    public long getSkip() {
         return skip;
     }
 
-    public int getPerPage() {
+    public long getPerPage() {
         return perPage;
     }
 }
